+++ packages/uql-orm/src/postgres/pgQuerierPool.test.ts
@@ -1,15 +1,10 @@
-import { types } from 'pg';
 import { AbstractSqlQuerierPoolIt } from '../querier/abstractSqlQuerierPool-test.js';
 import { createSpec } from '../test/index.js';
+import { configurePgNumericTypeParsers } from '../test/pgTypeParsers.util.js';
 import type { PgQuerier } from './pgQuerier.js';
 import { PgQuerierPool } from './pgQuerierPool.js';
 
-types.setTypeParser(types.builtins.INT8, (value: string) => Number.parseInt(value, 10));
-types.setTypeParser(types.builtins.FLOAT8, (value: string) => Number.parseFloat(value));
-types.setTypeParser(types.builtins.NUMERIC, (value: string) => Number.parseFloat(value));
-types.setTypeParser(types.builtins.INT8, (value: string) => Number.parseInt(value, 10));
-types.setTypeParser(types.builtins.FLOAT8, (value: string) => Number.parseFloat(value));
-types.setTypeParser(types.builtins.NUMERIC, (value: string) => Number.parseFloat(value));
+configurePgNumericTypeParsers();
 
 export class PostgresQuerierPoolIt extends AbstractSqlQuerierPoolIt<PgQuerier> {
   constructor() {